+++ src/main/java/com/example/concertreservationsystem/application/usecase/ConcertEventUseCase.java
@@ -15,36 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDate;
 
 @Component
-@RequiredArgsConstructor
-public class ConcertEventUseCase {
-
-    private final JpaConcertEventRepository concertEventRepository;
-    private final ConcertRepository concertRepository;
-    private final SeatInitializer seatInitializer;
-
-    @Transactional
-    public EventResponseDto registerConcertEvent(EventRequestDto requestDto) {
-
-        Concert concert = concertRepository.findByName(requestDto.getConcertName())
-                .orElseThrow(() -> new IllegalArgumentException("해당 콘서트를 찾을 수 없습니다."));
-
-        LocalDate eventDate = requestDto.getEventDate();
-        if (concertEventRepository.existsByConcertAndEventDate(concert, eventDate)) {
-            throw new IllegalArgumentException("이미 해당 날짜에 콘서트 이벤트가 존재합니다.");
-        }
-
-        ConcertEvent concertEvent = ConcertEvent.builder()
-                .eventDate(eventDate)
-                .totalSeats(requestDto.getTotalSeats())
-                .concert(concert)
-                .build();
-
-        concertEventRepository.save(concertEvent);
-        seatInitializer.initializeSeatsForEvent(concertEvent);
-
-        return new EventResponseDto(
-                concertEvent.getEventDate(),
-                concertEvent.getTotalSeats());
-    }
+public interface ConcertEventUseCase {
+    public EventResponseDto registerConcertEvent(EventRequestDto requestDto);
 }
