+++ src/main/java/com/example/concertreservationsystem/application/usecase/ConcertUseCase.java
@@ -9,25 +9,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 @Component
-@RequiredArgsConstructor
-public class ConcertUseCase {
+public interface ConcertUseCase {
+    public ConcertResponseDto registerConcert(ConcertRequestDto requestDto);
 
-    private final ConcertRepository concertRepository;
-
-    @Transactional
-    public ConcertResponseDto registerConcert(ConcertRequestDto requestDto) {
-
-        String concertName = requestDto.getName();
-        if(concertRepository.findByName(concertName).isPresent()) {
-            throw new IllegalArgumentException("중복된 이름을 가진 콘서트는 등록할 수 없습니다.");
-        }
-        Concert concert = Concert.builder()
-                .name(concertName)
-                .price(requestDto.getPrice())
-                .build();
-        concertRepository.save(concert);
-
-        return new ConcertResponseDto(concert.getName(), concert.getPrice());
-    }
 }
