+++ src/main/java/com/example/concertreservationsystem/domain/service/UserService.java
@@ -1,7 +1,6 @@
 package com.example.concertreservationsystem.domain.service;
 
 import com.example.concertreservationsystem.application.usecase.UserUseCase;
-import com.example.concertreservationsystem.domain.model.Concert;
 import com.example.concertreservationsystem.domain.model.User;
 import com.example.concertreservationsystem.domain.repo.UserRepository;
 import com.example.concertreservationsystem.web.dto.user.request.UserPointRequestDto;
@@ -10,22 +9,52 @@ import com.example.concertreservationsystem.web.dto.user.response.UserPointRespo
 import com.example.concertreservationsystem.web.dto.user.response.UserResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserService implements UserUseCase {
 
-    private final UserUseCase userUseCase;
+    private final UserRepository userRepository;
+    private final ReservationService reservationService;
 
+    @Override
+    @Transactional
     public UserResponseDto registerUser(UserRequestDto requestDto) {
-        return userUseCase.registerUser(requestDto);
+
+        String username = requestDto.getName();
+        if(userRepository.findByName(username).isPresent()) {
+            throw new IllegalArgumentException("이미 등록된 유저의 이름입니다.");
+        }
+
+        User user = User.builder()
+                .name(username)
+                .build();
+
+        userRepository.save(user);
+        return new UserResponseDto(user.getName(), user.getUuid());
     }
 
+
+    // 잔액 충전 (대기열 토큰 필수)
+    @Override
+    @Transactional
     public UserPointResponseDto chargePoint(String token, UserPointRequestDto requestDto) {
-        return userUseCase.chargePoint(token, requestDto);
+
+        // 대기열 토큰 검증
+        User user = reservationService.validateToken(token);
+
+        // 잔액 충전
+        user.addPoints(requestDto.getPoint());
+        userRepository.save(user);
+
+        return new UserPointResponseDto(user.getPoint());
     }
 
-    public UserPointResponseDto getUserPoint(String token){
-        return userUseCase.getUserPoint(token);
+    // 잔액 조회 (대기열 토큰 필수)
+    @Override
+    public UserPointResponseDto getUserPoint(String token) {
+        User user = reservationService.validateToken(token);
+        return new UserPointResponseDto(user.getPoint());
     }
 }
