+++ src/main/java/com/example/concertreservationsystem/web/controller/queue/QueueController.java
@@ -1,7 +1,7 @@
 package com.example.concertreservationsystem.web.controller.queue;
 
 
-import com.example.concertreservationsystem.domain.service.QueueService;
+import com.example.concertreservationsystem.application.usecase.QueueUseCase;
 import com.example.concertreservationsystem.web.dto.queue.response.QueueResponseToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -15,11 +15,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/queues")
 public class QueueController {
 
-    private final QueueService queueService;
+    private final QueueUseCase queueUseCase;
 
     @PostMapping
     public ResponseEntity<QueueResponseToken> addQueueToUser(@RequestParam String uuid) {
-        QueueResponseToken responseToken = queueService.addQueueToUser(uuid);
+        QueueResponseToken responseToken = queueUseCase.addQueueToUser(uuid);
         return ResponseEntity.ok(responseToken);
     }
 }
