+++ src/main/java/com/example/concertreservationsystem/web/controller/reservation/ReservationController.java
@@ -1,6 +1,6 @@
 package com.example.concertreservationsystem.web.controller.reservation;
 
-import com.example.concertreservationsystem.domain.service.ReservationService;
+import com.example.concertreservationsystem.application.usecase.ReservationUseCase;
 import com.example.concertreservationsystem.web.dto.reservation.request.ReservationRequestDto;
 import com.example.concertreservationsystem.web.dto.reservation.response.ReservationResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/reservations")
 public class ReservationController {
 
-    private final ReservationService reservationService;
+    private final ReservationUseCase reservationUseCase;
 
 
     /**
@@ -25,7 +25,7 @@ public class ReservationController {
     public ResponseEntity<ReservationResponseDto> rvConcertToUser(@PathVariable Long concertId,
                                                                   @RequestParam String token,
                                                                   @RequestBody ReservationRequestDto requestDto) {
-        ReservationResponseDto responseDto = reservationService.rvConcertToUser(concertId, token, requestDto);
+        ReservationResponseDto responseDto = reservationUseCase.rvConcertToUser(concertId, token, requestDto);
         return ResponseEntity.ok(responseDto);
     }
 }
